backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic-settings==2.5.0

# Document Parsing
PyMuPDF>=1.24.0
python-docx==1.1.0
chardet==5.2.0

# Chunking & Tokenization
nltk==3.9.0

# Embeddings
sentence-transformers==3.0.0

# Vector Store
chromadb==0.5.0

# LLM
openai==1.50.0           
anthropic==0.34.0     
google-generativeai>=0.7.0    

# Utilities
python-dotenv==1.0.1
numpy>=1.26.0
